Suppression de tous les warnings -Wswitch-bool (53 occurrences)
- faders_core_24.cpp : factorisation 3x bloc getRepeat() → player_toggle_loop(n) + if/else pour les switch(bool) restants
- audio_core.cpp / audio_core.h : ajout player_toggle_loop(int n)
- keyboard_functions2.cpp : cast (int) sur ClassicalChannelView et index_do_hipass
- logicals_intres.cpp : cast (int) sur index_blind, wizard_amplitude_is_global, multiple_direct_chan, index_grid_tostep_or_not, channel_view_mode_builder
- saves_export_pdf2.cpp : cast (int) sur 10 variables bool en switch
- gui_boutons_rebuild1.cpp : switch(index_show_main_menu) → if/else
- faders_operations.cpp, trichro_visu2.cpp : cast (int)

+++ Src/faders_operations.cpp
@@ -524,7 +524,7 @@ numeric_postext=0;
 if(last_ch_selected>0 && last_ch_selected<513)
 {
 ClearDock(fad,dk);
-switch(multiple_direct_chan)
+switch((int)multiple_direct_chan)
 {
 case 0:
 DockTypeIs[fad][dk]=10;//direct channel mode
